requirements.txt
allure-pytest              2.15.0
allure-python-commons      2.15.0
attrs                      25.3.0
certifi                    2025.7.14
cffi                       1.17.1
charset-normalizer         3.4.2
chromedriver-autoinstaller 0.6.4
chromedriver-binary        141.0.7358.0.0
colorama                   0.4.6
cryptography               45.0.6
execnet                    2.1.1
flake8                     7.2.0
greenlet                   3.2.4
h11                        0.16.0
idna                       3.10
iniconfig                  2.1.0
Jinja2                     3.1.6
MarkupSafe                 3.0.2
mccabe                     0.7.0
outcome                    1.3.0.post0
packaging                  25.0
pip                        25.2
pluggy                     1.6.0
psycopg2                   2.9.10
psycopg2-binary            2.9.10
py                         1.11.0
pycodestyle                2.13.0
pyOpenSSL                  25.1.0
pycparser                  2.22
pyflakes                   3.3.2
Pygments                   2.19.1
PySocks                    1.7.1
pytest                     7.4.0
pytest-forked              1.6.0
pytest-html                4.1.1
pytest-metadata            3.1.1
pytest-xdist               2.5.0
python-dotenv              1.1.1
PyYAML                     6.0.1
requests                   2.31.0
selenium                   4.9.1
six                        1.17.0
sniffio                    1.3.1
sortedcontainers           2.4.0
SQLAlchemy                 2.0.43
tqdm                       4.67.1
trio                       0.30.0
trio-websocket             0.12.2
typing_extensions          4.14.1
urllib3                    2.5.0
urllib3-secure-extra       0.1.0
webdriver-manager          4.0.2
websocket-client           1.8.0
wsproto                    1.2.0